+++ CHANGELOG.md
@@ -3,8 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
 ## [Unreleased]
+
+## [5.1.0] - 2025-08-12
+### Added
+- Add support for token-based pagination in WorkspaceResources.ListWorkspaces()
+- New TokenPaginationParameters class to support paginationType, lastKey and maxItems parameters
+- New TokenPaginatedResult<T> class for token-based pagination responses with data and lastKey properties
 
 ## [3.0.0] - 2022-12-07
 ### Updated